Create week-06-murray.md
In my opinion, I believe Murray's predictions are somewhat accurate. I don't think that they will be 100% true though, as I feel as if some of his predictions are far fetched. I don't believe it will be the new form of "serial drama", which I'm not completely sure what was meant by this. I do, however believe, that it's true when he says that virtual environments are an extensional of the fictional world. This is present now in ways such as simulations, as well as social interactions in my opinion due to them all being associated with real life encounters. 

The Mandalorian uses multiple forms of digital technology in order to achieve the effect it achieves. According to further research, this is all thanks to pre production  work. Lots of time has been spent on VFX workflows well before the principal photography deadline, according to the ibc.org article "Behind The Scenes: The Mandalorian's Groundbreaking Virtual Production. It relates very much so to what Murray was referring to in the way that their stages are nearly all very technical, leaning more into a virtual reality than anything else. VFX, Parallax, and enhanced LED lighting also are present on this set. Parallax, which according to the aforementioned article, as a camera effect where the camera is moving rapidly alongside a set as a character moves, in order to make it seem as if you're following right along with them. This adds a more intense, dramatic affect to the film. Lastly, LED lights, which also include the reflection, refraction, and enhancing of specific color lights, also add a virtual affect to the set of the Mandalorian. As stated in the article, this creates special affects such as sun glares, light, and even darkness as if it were actually happening right there and then in the moment, in real life.